Add an optional timeout parameter to inotify_wait
Optionally timeout inotify_wait to update qbittorrent's listening port
on a periodic basis. The default value (if the environment variable is
not set) is to have no timeout, so there is no change in functionality
if the parameter is unset.

This helps with situations where, for example, qbittorrent has restarted
and is using its default listening port, but gluetun has not updated
its forwarded port, leaving qbittorrent unaware of the forwarded
port.